Move attire outside of QueuedAction & refundTime for actions

+++ contracts/debug/DebugStorageSlots.sol
@@ -97,7 +97,7 @@ contract DebugStorageSlots {
     uint16 actionId;
     uint40 startTime;
     uint24 elapsedTime;
-    uint64 queueId;
+    uint40 queueId;
     // Boosts active at the time this was generated
     BoostType boostType;
     uint16 boostValue; // Varies, could be the % increase
@@ -165,7 +165,7 @@ contract DebugStorageSlots {
     uint16 leftHandEquipmentTokenId; // Shield, can be empty
     uint24 timespan; // How long to queue the action for
     CombatStyle combatStyle; // specific style of combat,  can also be used
-    uint64 queueId;
+    uint40 queueId;
   }
 
   struct QueuedAction {
@@ -182,7 +182,7 @@ contract DebugStorageSlots {
     CombatStyle combatStyle; // specific style of combat,  can also be used
     uint40 startTime; // When the queued action started
     bool isValid; // If we still have the item, TODO: Not used yet
-    uint64 queueId; // id of this queued action
+    uint40 queueId; // id of this queued action
   }
 
   struct ActionInfo {
